Make default action in embark occur buffer for imenu work
This involves popping to the buffer imenu was called from. I added
that as a buffer-local pre-action hook in the occur buffer when called
from imenu.

This phenomenon of needing some special setup in the pre-action hook
might happen for other commands besides imenu. If I come across
another example, I should implement a general mechanism for it. Maybe
an embark-occur-pre-actions-hooks-alist.