Rename main commands to embark-act-noexit, embark-act
The commands embark-act and embark-exit-and-act differ only when
called from the minibuffer: embark-act does not exit, the other one
does.

@clemera brought up the point in the discussion of PR #24 that
embark-exit-and-act is misleading because it doesn't actually exit
anything unless called from a minibuffer. After considering many, many
aleternatives, I noticed that renaming embark-act to embark-act-noexit
and embark-exit-and-act to embark-act solves the problem: it is
always correct to say the new embark-act-noexit does not exit!