Run embark-post-action-hook run after the action is really finished!
With the previous implementation of embark--cleanup the
embark-post-action-hook ran soon after the target was consumed. This
meant it often ran while the the action was still in progress! If the
action allowed editing the target or if it prompted for a second
argument, then the hooks would run before the action concluded.

This kept the embark-post-action-hook from being used, for instance,
to refresh an Embark Live Occur buffer after an action --in case the
candidate list changed as it does with rename-file or kill-buffer.
(Note that to refresh Icomplete or Selectrum, embark-pre-action-hook
is enough, but it is a little trickier --but possible-- to use that
hook to refresh an Embark Live Occur buffer.)

This new implementation has other advantages: since the cleanup runs
right after the next command after embark-act, we can catch the case
when said command is not an action to consume the target and produce
an error message. This makes remapping self-insert-command to
embark-undefined unecessary too.