Change string passed to completion-metadata
I don't think this really matters: while in theory the
completion-metadata can depend on the input string and even on the
location of the point, I don't think it ever does in completion tables
that arise in practice.

But anyway, this is supposed to be more correct than sending the
entire minibuffer contents. Plus, it also works better for Selectrum,
as it excludes the candidate display.