Postpone cleanup for embark-act-on-region-contents
I hope this is the last bug from the cleanup reimplementation!

If you use embark-act-on-region-contents as the action on an active
region, it runs embark-act for you with target the contents of the
active region. Cleanup was running right after
embark-act-on-region-contents, this made embark forget the
classification of the region contents by the time the user selected
the action.