Pop to target buffer for default action
Previously popping to the target buffer was done for all actions if
the originating command was imenu, or if the completion category was
line (this was for consult-line and consult-outline).

But thinking about it some more, I believe it should always be done
for the default action, but never for the other actions (unless they
want do it themselves; embark-insert is almost an example: it inserts
in the target buffer, but doesn't pop there---maybe it should!).

With this change, embark-occur should work better for
consult-{line,outline,mark}, with out any special-casing anymore!

It's also fun now to use embark-occur on M-x, and try out a bunch of
commands in the buffer you started with. :)