Remove the last of the annotators
Now all annotator logic has been removed from Embark.

This involved removing support for custom formats for
embark-keymap-help, since that was doing some annotation of its
own (and not was it annotating, it also included some logic for
configuring the display of annotations, namely for truncation.)

Now, instead, embark-keymap-help reports the command metadata category
when it prompts for a command, so other mechanisms can handle the
annotation. (Sorry @clemera, but I think this is for the best.)

One slight change in behavior is that now the key bindings in
embark-keymap-help are just displayed, and not physically part of the
completion candidates (this is so the command annotator can recieve
just the command name).