embark-completing-read --> embark-live-occur-after-{delay,input}
Remove embark-completing-read. In its place there are two new
functions:

- embark-live-occur-after-delay, and
- embark-live-occur-after-input

These are meant to be added to minibuffer-setup-hook. The -delay
variant behaves exactly as the old embark-completing-read function.
The -input variant waits for the first change in the minibuffer, then
sets off a timer of duration embark-live-occur-initial-delay and then
pops up an Embark Live Occur buffer.