General fix for mouse menu popups and dialog boxes with embark-act
In issue #39 a situation is described in which a file dialog appears
when using embark-act. I fixed that by binding use-dialog-box to nil.
But now I noticed the bookmark-jump-other-{window/frame} would pop up
a mouse menu, because last-nonmenu-event is nil and those commands
check for mouse usage with listp... (I think they really meant to
check for non-nil lists.)

A general solution to both problems is to pretend the commands where
run by pressing RET, which is what this fix implements.