First attempt at implementing the simplified action system
The trouble with the old system was that the action was run by the
event loop, not under the control of Embark. This led to a complicated
choreography to ensure (unsuccesfully, at times) that all the
necessary cleanup happened in a timely fashion. In particular, I never
really got the old implementation to come back to the starting buffer
in all cases after running the (possibly canceled) action in the
target buffer. Let's hope that works better now.

The idea in this new implementation is to directly run the user's
commands under control of embark-act and embark-become. I'm sure there
are plenty of kinks to iron out, but at first blush this seems much
simpler.

It looks like I'll also be able to remove many of the buffer-local
variables.