Fix default-directory for annotators in Embark occur list view
This is a little tricky: for key bindings we want the annotator to run
in the target window, so we get those bindings instead of whatever is
active in the Embark occur buffer; on the other hand, for Marginalia's
file annotations, we want the default-directory to be that of the
Embark occur buffer, independently of where it came from. So we do run
annotators in the target window but we smuggle in the
default-directory of the Embark occur buffer.