Unify target acquisition with classification!
Most often when you find a target, say the url at point, you already
know its type (url, in this example), so it makes sense to do target
acquisition and classification at the same time.

Now the functions in embark-target-finders and
embark-candidate-collectors must return a pair of the form
(cons type target) and a list of the form (cons type candidates),
respectively.

The variables embark-classifiers and embark-target-classifiers are now
gone. Also gone is the misguided embark-input-getters variable used by
embark-become.

A casualty of this restructuring is embark-act-on-region-contents.
That action would attempt to classify the contents of the region and
offer the appropriate action keymap. I don't think this will be
missed, very rarely does the region contain solely a file name or a
symbol, etc. And if it does, it is often the thing at point, for which
actions would be available by *not* marking the region.