Refine virtual-buffers to real type; strip number prefix from lines
This fixes #74.

We are adding functionality for Consult without introducing any
dependency on it. The price for that is that if Consult changes its
format for virtual-buffers or for lines, this needs to be updated.

Instead of just special-casing these two transformations I could add a
general mechanism for registering transformations. That change is very
easy, but in the interest of not having more public API than
necessary, I won't do it until other useful transformers are
discovered.

The virtual-buffer change is particularly cool: now both
embark-act(-noexit) and embark-(live-)occur will offer you appropriate
actions for each candidate! Sadly
embark-occur-direct-action-minor-mode will only bind the general
actions in an Embark Occur buffer for virtual-buffers. (Does anyone
even use the embark-occur-direct-action-minor-mode?)