Simplify the way embark live occur buffers are made ephemeral
I used to have an artificial call to embark-occur--kill-live-occur-buffer
in the middle of embark-occur-noselect, because even though I thought
I was killing the live occur buffers in the minibuffer-exit-hook it
"didn't always run". I now understand what was happening: the
minibuffer-exit-hook actually was running but *after* killing the
local variables so that embark-occur-linked-buffer was already nil by
the time I tried to kill the buffer. By instead putting a closure on
the exit hook that kills the correct buffer I don't need the call to
embark-occur--kill-live-occur-buffer anymore, and indeed, I don't need
that function anymore. (I won't remove it right away because I know
@protesilaos uses it.)

This change is in preparation for #44.