Transform minor mode lighters to the mode name (fix #75)
This is as much of a fix for #75 as I will probably
implement. Since, as @minad explained to me in #75, the lighter for
some modes such as flycheck is not constant and so the lookup might
fail depending on which buffer you run it in. I'll ignore that problem
for now.