Fix #84: Deal with unprefixed virtual-buffers
I had incorrectly virtual-buffers targets would always contain the
preix typing character, but of course, if you type something that
doesn't actually match any virtual character the minibuffer contents
itself will be the target. I was incorrectly stripping the first
character and classifying them as general. Since consult-buffer will
create a new buffer in the non-matching input case, it is better to
classify them as 'buffer (and, of course, to not strip the first
character!).