Reform information caching strategy to fix #79
Now the only information cached for a minibuffer is the command that
opened it, which embark-default-action needs to know. This is done in
an autoload, so that embark-default-action can work even the first you
use it.

A nice side effect of the reform is that now the candidate collector
is only run once for embark-occur. Embarrasingly it was previously
being run twice, and one of the two times the candidates were being
discarded and only the type kept!

Also, now for *Completions* buffers the type is the minibuffer
completion category or nil if the completion-reference-buffer is not a
minibuffer. Embark does not gather the completions shown in
the *Completions* buffer, Emacs does, so it makes no sense to run the
candidate collectors and, assuming they would magically produce the
list of completions, trust the type they returned. This is what we
previously did! And of course, it works with the default value of
embark-candidate-collectors but the user could change that! So
instead, since the *Completions* are not under Embark's control, we
make the best guess at which type of actions would be most useful for
them, which for *Completions* coming from a minibuffer is the
completion category.