Move first revert of embark occur buffer earlier
The recent information cache reform, means the type of the candidates
is no longer save by embark--cache-info, but rather at time of first
revert, when the candidate collector is run (which is great: that is
the earliest the candidates need to be collected, and we find out
their type at the same time). But then we need the first revert to
happen before embark-occur severs the link to the minibuffer and
exits.