Revert "Initial implementation for embark-repeat (see #85)"
This reverts commit 622f6eabfce4e3f15a34774646f2a9325b6c852f.

The embark-repeat command was an experiment that failed because
"minibuffer contents during the minibuffer exit hook" is not a good
definition of "last input", since pressing RET can modify the
minibuffer contents to match the selected candidate.

As discussed in #85, we have decided to remove this experiment in
favor of a better behaved and more featureful solution that will be
made available as a separate package.