Set this-command to the command we are becoming (fix #93)
When you take over the command loop, you have to remember the command
loop does plenty of things besides just running the next command, and
it's easy to forget emulating something the command loop does for you.

Here I forgot to set this-command to command we are becoming. Embark
relies on this-command being set correctly at the time of first
minibuffer prompt.