Initial implementation of occur mode exporter for consult-location
This fixes #98 for commands like consult-line and consult-mark where
all the lines come from the same buffer.

For consult-global-mark where the lines come from many buffers there
is still a little work to do, we need to group the one that come from
the same buffer and make a little header line for each buffer. Other
than that, it actually works already: each line takes you to the
correct buffer and even occur-edit-mode works!