Add a few new actions for commands and for variables
The default configuration now has new action keymaps for commands and
for variables, which have the symbol action map as their parent. For
commands, we have M-x, global-set-key and local-set-key actions. For
variables, we have set-variable, customize-set-variable and
customize-variable actions. (The last one creates a Customize buffer,
while the one before last uses the minibuffer.) Also, the binding for
Info-goto-emacs-command-node has been moved to the command map.