Put everything under avy-embark-collect prefix
Previously the avy-embark-collect--choose function was named
avy-action-choose because I thought that to get the nice help message
that appears when you press ? in avy the actions needed to start with
"avy-action-". This is not true, the first (length "avy-action-") = 11
characters of the function named are removed but it is not checked
that those 11 characters are "avy-action-". So I can call the function
avy-embark-collect--choose and the help message will say
"collect--choose", which is not too bad.

It's lucky that "action" and "embark" have the same number of letters.