Fix recursive embark-act inside recursive non-embark minibuffers
If we dynamically let-bind a variable around read-from-minibuffer, heavy
recursive minibuffer users will stumble upon various problems. For example, an
inner embark-act will always throw to an outer embark act, even if there are
multiple non-embark minibuffer levels in between.

To fix this, we use setq-local and a modified version of
minibuffer-with-setup-hook, that doesn't self-destruct, so it can work for
multiple minibuffers. But we also make sure it works only for minibuffers whose
level is directly under the current one.