Add easy way to override or provide fallback for default action
For targets that come from the minibuffer, the default action is the
command that opened the minibuffer. This can now be overridden on a
per category basis with embark-default-action-overrides. For target
that don't come from the minibuffer, there used to be no default
action. One can now be specified in embark-default-action-fallbacks.