embark-goto-location use directory of buffer action was called from
Actions are run in the target window, so in this case whatever window
was selected when the command that produced the xref-location
candidates ran. In particular, we inherit the default-directory of the
buffer in that window, but we really want the default-directory of the
minibuffer or collect window we call the action from, which is the
previous window, since the location is given relative to that
directory.

This is an attempt to fix #149.