Temporarily mark window where action originates as dedicated
This is meant to fix #157.

I think that between the change suggested by @jakanakaevangeli in the
discussion of #152 (namely, not restoring the originating window when
an action selected a window different from the action window) and this
change (which marks the window where the action is run as dedicated
for the duration of the action), we have fairly reasonable window
behavior for all common actions. Further experience will tell if this
is really true. It may still be necessary to make the window behavior
configurable, but I hope we can get away with hardwired behavior
(although maybe the current hardwired behavior still needs some
tweaking).