When the default action is embark--command, run it on original target
This fixes #180.

Background: the default action is given, in order of priority, by:
(a): embark-default-action-overrides, (b) embark--command, (c) the
binding of RET in the relevant action keymap.

This change means that when the default action is given by
embark--command, then it recieves as argument the untransformed
target. This is meant for commands that prompt in the minibuffer for
candidates that are somehow "mangled", like several Consult commands
which prefix their candidates with unicode tofu. The tofu is necessary
for that command but should be removed for other actions.