Change way we call Marginalia annotators (fix #185)
With this change Marginalia annotators are only used when
marginalia-mode is active, which might be considered more respectful
of users's preferences.

It also has the feature that annotators run under the
marginalia--context macro, which probably has no effect in this case.

Also, now if the mechanism Marginalia uses to augment metadata changes
this code will need to be updated.