Add embark-prefix-help-command
This function is intended to be used as a value for
prefix-help-command. That configuration makes C-h after a prefix
prompt you for a command to run using completion. The keybindings are
displayed with the commands (if your completion UI displays completion
candidates, of course), and you can type @ to use the key bindings.

See discussion in raxod502/selectrum#504.