Safer guess for the default-directory
As discovered by @hmelman in (the unrelated) #235, when using selectrum with prescient and keeping completion-styles at the default value it is very easy for the minibuffer contents to match according to prescient but not according to the completion-styles. When that happens, the base-position computed by
(cdr (last (completion-all-completions ...))) is at the end of the input and thus the guessed default-directory includes some gibberish at the end.

To fix this, when there are no matching candidates, we just look for the last slash in the minibuffer contents. This is not perfect, but should cover most cases.