Alternative implementation for defun and sexp targets and actions
The main impetus here was to get the basic functionality working in
non-lisp programming modes. I'm not sure it was worth the
complication. Here are some notes:

1. Most of the actions in the defun map really are Emacs Lisp
specific. Only marking and narrowing can work in other modes, plus, of
course the inherited sexp manipulation commands that are textual:
indent and kill.

2. At least I got indent and kill to work correctly for thingatpt
defuns. The implementation uses indent-region and kill-region. The
reason for this is that a defun is not a sexp in many modes, and also
indent-sexp doesn't indent according to the mode specific rules.

3. In C++ thingatpt disagrees slightly the -defun commands:
mark-defun, beginning-of-defun, etc. Both mark standalone functions
and class methods, but thingatpt does not consider the entire class
definition to be a defun. This is not what happens in java-mode, there
thingatpt and the -defun commands seem to agree, and in particular,
agree that a whole class is a defun.