Customizable structure for verbose indicator buffer
We introduce the customizable list
embark-verbose-indicator-buffer-sections, which denotes the sections
to be included in the indicator buffer, in the user's desired order.
Besides symbols denoting specific sections (target, other-targets,
etc.), we allow the symbol newline and literal strings, which are
inserted, well, literally.

For each of the elements in the list there's a corresponding insertion
function, which users can advice to provide even more control on how
the sections are printed (we could define customization variables for
each inserter, but that seems to be, perhaps, too much).

The default value of embark-verbose-indicator-buffer-sections
reproduces the current structure.

As another example, i prefer the title to be below the
bindings (because i use bottom placement), put together the current
target and the alternative ones in the same line (i've got plenty of
horizontal space) and omit the cycle key indicator (i don't need that
reminder), so i've got

   '(bindings newline target "  " other-targets)

as the value of embark-verbose-indicator-buffer-sections.

Yes, newline is redundant (one can use the literal string "\n" with
the same effect), but it looks nicer in the customization buffer as a
symbolic choice.

An alternative that might be worth considering is using formatting
functions returning strings instead of inserters which write in the
buffer, specially if one wants to make them customizable (although a
bit of flexibility is lost that way if not).