Simplify indicator logic, update a couple of spot to new protocol
embark-act was a little more complicated than necessary and was
missing indicator cleanup in the case the action was cancelled.

I also decided to make embark--prompt resposible for calling the
indicator closure, so as to not duplicate this in embark-become and
embark-act.

Finally, embark-become and the inner-indicator were not calling the
correct closure for clean up.