Use strings for regions, use pre-action hooks for sexp & region
Remove the special casing for region targets. Now they return the
region contents and bounds like normal target finders.

The write-region and append-to-file region actions prompt for a file
name and the region should not be injected there. To handle this I
added a new pre-action hook that ignores the target for those
commands.

I also decided to remove the sexp action wrappers and instead use a
pre-action hook to position point at the beginning of the sexp.

The only change in the embark--act logic required for this was to run
the pre-action hook *after* setting this-command to the action. I had
always intended for pre-action hooks to be able to find out what the
action was, but had mistakenly run the hook slightly too early.