If an action is directly bound, give it no special treatment
This fixes #335.

Several commands have special treatment in the embark-keymap-prompter,
including execute-extended-command. When you select that action, the
keymap prompter assumes you didn't mean it, that what you really want
is to choose an action at the M-x prompt.

With this change, if you choose an action through a key bound in the
action keymap, it receives no special treatment and is used as the
action.

In particular, now for execute-extended-command it matters how you
selected it: if you selected it through a binding in the action
keymap, then it does get run as an action; if instead you selected it
via a global key binding such as M-x, you then get the special
treatment of being prompted for an action.

Thanks to this change we now longer need the embark-execute-command
alias to execute-extended-command, and can bind x to
execute-extended-command directly in the command action map.