Add binding for info-lookup-symbol in identifier map
info-lookup-symbol is pre-configured to look up symbols in info
manuals appropriate to the major mode of buffer the symbol is in. For
example, in sh-mode it uses manuals for bash, coreutils, diff and
gawk.

We still need the emacs-lisp specific function
embark-info-lookup-symbol because we detect Emacs Lisp symbols in all
non-prog-mode buffers and info-lookup-symbol only uses emacs-lisp
manuals for some of them (help-mode, Custom-mode, apropos-mode).