Only start at region targets if explicitly using mark action
As u/divinedominion reported on reddit, if you have an active region, run embark-act, and use a repeatable action, the next time you are offered the region target instead of the same type of target the repeatable action used.

With this change the target cycle only switches to starting at the region target after explicitly using the 'mark action, not if the reason there is an active region is that there was one before the first embark-act.