Remove consecutive duplicated targets
This was proposed and implemented by @minad but later we decided to
get rid of the deduplication as unlikely. But it is not unlikely! It
happens all the time when acting in a *Completions* or Embark Collect
buffer you get from describe-{function,variable,etc.}: the symbol at
point target and the completion at point target are exactly the same!