Make embark--quit-and-run = apply outside a minibuffer
embark-export calls embark--quit-and-run from both the minibuffer and
other buffers, which produces a harmless error in the non-minibuffer
case. To get rid of the error I had two choices: handle the
non-minibuffer case in embark-export or in embark--quit-and-run.
Having embark--quit-and-run handle the non-minibuffer case simplifies
using it a tiny bit in a few places, so I went with that.