Set occur-highlight-regexp to regexp that matches the whole line
The docstring for occur-highlight-regexp says occur will highlight
entire line upon pressing RET on a search result if the variable
occur-highlight-regexp = t, but apparently this exporter is somehow
sidestepping some machinery that makes this happen and I see an error
message instead. So let's set it to a regexp that matches the entire
line to get the highlighting.

(The entire line is the only reasonable thing to do here: this
exporter is used for consult-line' where a completion style can match
anyway it wants to; and for consult-outline, where really the entire
line is the match.)