Demote user-errors for repeating actions
Earlier we had special cased the next/previous symbol commands, by
writing custom commands that printed a message instead of signaling a
user-error as would be customary.

Now with some usage experience, I think that (1) forward-sexp and the
like all deserve this treatment and (2) it would be too much of a
hassle to have to write these wrappers for them. The best course of
action is to demote user-errors (but not other types of error!) when
running a repeating action.

I've also change the next/previous symbol commands to use the more
idiomatic user-error.