Make argument of embark-find-definition a string (fix #357)
Apparently in very recent Emacs 28's xref-find-definitions only takes
string arguments (before it could handle both symbols and strings). I
had made the argument to embark-find-definition a symbol because I
thought doing so provided completion from the obarray, but this is not
the case. Given that, there is no disadvantage in using a string
argument.