Replace `embark--quit-and-run` with `embark--run-and-quit`
(Follow up for #362)

I expect this to work well for all kinds of actions, since actions generally
work for embark-quit-after-action=nil. This has the advantage that the actions
are executed within the current dynamic scope, where for example
default-directory is bound.

The only difficulty here is to properly restore the window configuration. I
suspect the minibuffer restoration of the window configuration is the only
reason why we used embark--quit-and-run in the first place? There may still be
bugs!

We should really consider setting embark-quit-after-action=nil by default since
it is the better behavior and given the dynamic scoping issue it is also more
natural. I argued before that the actions should be executed in the original
window which hints at closing the minibuffer first. But since we always switch
to the original window before executing the action anyways in order to support
non-quitting actions we are good.

There is one difference in behavior however. Since we always execute the actions
within a recursive minibuffer now, quitting from the action for
embark-allow-edit-actions will quit to the outer completion command. Can we live
with that? Maybe this behavior is even reasonable since aborting the action will
take as back to where we came from. Alternatively we can catch the quit
exception and quit twice, further complicating the quitting logic...