Use thing-at-point for URL to avoid potential bound error
Previously we go the URL at point from ffap-url-at-point but got the
bounds from thing-at-point. This led to a potential mismatch between
the target and the highlighting. This changes uses thing-at-point for
both operations, thus fixing the problem.

The reason for using ffap-url-at-point was that it also recognized
email address, even when not prefixed by "mailto:" (and browse-url by
default handles mailto URLs by composing a new email in Emacs, which
is nice to have!). For email addresses (without the "mailto:" prefix),
the potential mismatch did indeed occur and the email address was not
higlighted by embark-highlight-indicator. With this change we still
recognize email addresses but now also correctly highlight them.