Add an Emacs Lisp library file target with lower priority than symbols
We already have an Emacs Lisp library target: if a symbol is not a
function, variable or face, and is a library name, then it gets
treated as a library. See embark-library-map for library actions.

This change adds a second target for libraries that treats them as
files and places is right after the identifier target, so that cycling
once from a library target switches to considering it as just a file.
This can be useful for copying the file path; deleting, copying or
renaming the library file, etc.