Make file targets consistently full paths
Previously file completion in the minibuffer produced full paths, but
acting on file targets in other buffers varied: in dired you get
relative paths, for paths explicitly written out you got them as is
whether relative or not, and for files "guessed" by ffap you got full
paths. We have the actions embark-save-relative-path and
embark-insert-relative-path to be able to get the relative path, but
sometimes there was no way to get the full path.

With this change file targets should all be full paths, so that the
kill-new (w) and embark-insert (i) actions consistently deal with full
paths, and of course we still have the relative-path versions of those
actions bound to W and I.

This address a concern in issue #376.