Support typing help-char in embark-keymap-prompter
When the user types C-h (the default value of help-char) after a
prefix key, emacs shows help on the prefix keymap by calling
prefix-help-command.  This change makes the embark keymap prompter
behave similarly: it switches to the completing-read prompter for the
prefix keymap typed so far.

This is in addition to the existing functionality, where C-h is bound
to embark-keymap-help in embark-meta-map. That is only useful for
getting help on the embark action map, not for other keys.