Add new email target category
Emails can occur as a "mailto:" URL or on their own. Previously
embark-target-url-at-point would turn any lone email address into a
mailto URL. One would compose a mail by using the browse-url action,
but this is a bit fragile since it depends on the mailto entry of
browse-url-default-handlers not being customized or overridden in
browse-url-handlers. Also, people usually just don't want the
"mailto:" prefix if they save an email address on the kill-ring.

So, now, instead we turn both mailto URLs and standalone emails to a
plain unprefixed address and have a dedicated embark-compose-mail
action, which is a trivial wrapper around compose-mail (which we would
have been able to use directly if it had a different interactive
specification).

Left for future work: adding other sensible actions to emails. Perhaps
an action to add the address to a contact list? The problem with that
one is that there are so many contact books one might use in Emacs:
ecomplete, bbdb, ebdb, eudc, org-contacts, ...! (That sounds like a
good candidate for a wiki entry, though.)