Add embark-help-key, remove embark-keymap-help from embark-meta-map
Treat the help key specially such that it is always available when
acting with the keymap prompter, even if the action keymap does
not contain such a binding. This special treatment is needed to
ensure that the Embark help is always available. It is possible
to disable the help binding by setting embark-help-key=nil. The
embark-cycle-key has a similar treatment and is always available.