Have embark-insert insert some whitespace if needed
For a while I've been thinking that embark-insert should put some
space around the string it inserts to avoid having it run together
with the text in the buffer. The lack of padding made it very awkward
to insert several string in a row. For example, I often write comment
for GitHub or Reddit about Emacs in a Markdown buffer and use C-h f to
collect the function names I want to mention. Inserting more than one
of them in a single C-h f session made them run together. With the new
embark-act-all command this problem is even worse!

It seems unlikely that some people may prefer the old mash-em-up
behavior, so I'm not initially making this configurable. If a user is
happy with the padding but for some specific use case wants no spaces
it is easy to get to the space with motion commands and delete them.
By contrast, with the old behavior it was unnecessarily hard to insert
the missing spaces since you could't use easy word navigation commands
to get to the right spot: it required moving by characters or using
isearch or clicking with the mouse, etc.