Don't inject target into pre-action hook minibuffers!
I can't believe this bug was undiscovered for so long. The minibuffer
injection applied to the pre-action hook if it happened to open a
minibuffer! This was only discovered due to the new embark--confirm
action which uses the minibuffer or not depending on the value of the
variable y-or-n-p-use-read-key.

This change also fixes #439.