Use "for display" versions of candidates in Embark Collect buffers
We now flatten display properties and remove invisible portions. This
is done because tabulated list, when truncating strings to the given
column width, ignores any display properties and invisible portions of
strings. So, for example if a string has 5 visible characters, then
100 invisible one, and then another 5 visible characters and the
column width is 15, tabulated list mode will take the first 15
characters of which only the first 5 are visible! This leads to
unexpected ellipsis and seemingly missing information in Embark
Collect buffers, as reported in bdarcus/citar#504.