Replace embark-allow-edit-actions with a setup hook
The variable embark-allow-edit-actions is unnecessary, since its
effect can be achieved with just a setup hook! That's what the new
embark--allow-edit function is for.

While this change is a nice simplification anyway, it's really
prompted by a bug I introduced in commit c672dcf1eb8c1be0462778b2ab5a115150568c83. :P

There I noticed that the minibuffer-setup-hook used for injection was
also applied during the pre-action hook! That meant that if a
pre-action hook happened to prompt the user in the minibuffer, the
target would be erroneously injected there. This caused problems with
embark--confirm, so I tightened the scope of the injection to exclude
the pre-action hook and just apply to the action. Unfortunately, that
completely broke the implementation of embark--ignore-target as a
pre-action hook, which worked by artificially reading from the
minibuffer to gobble up the target.

Now, I want to keep the narrower scope of the target injection, since
it makes more sense, so to fix that bug I decided to reimplement
embark--ingore-target as a setup hook (which also makes more sense
since setup hooks are exactly about what happens during target
injection). This required swapping the order of calling the pre-action
hook and scheduling the minibuffer exit.

Once I did that, I noticed that now embark--allow-edit was possible as
a setup hook and decided to remove the embark-allow-edit-actions
variable. The fewer variables the better, and we're already committed
to extensive hook use.