Change wording of message in embark-completing-read-prompter
This changes the text of the user warning given in the completing-read
prompter in case embark-cycle-key = nil, there are no further targets
and the user presses the embark-act key binding.

It previously said: "Single target; can't cycle. Press `%s' again to
act." But embark-completing-read-prompter is not used only for help
during embark-act: it is also used for embark-prefix-help-command,
embark-bindings-in-keymap, embark-bindings-at-point, and
embark-bindings; and in those cases the "single target" part of the
message was inaccurate.

I noticed this while fixing #450.