Remove use of delete-blank-lines in embark-insert (fix #451)
The call to delete-blank-lines was there to deal with the case of
insert a multiline string that ends in newline (a common case for me
in which this arises is using embark-insert right after the closing
parenthesis of a defun to make a copy of it). But I didn't realize
that delete-blank-lines also deletes the blank lines below the current
when the current line is not blank!